travel_backend/requirements.txt
Django>=4.2.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.2.0
django-cors-headers>=4.0.0
django-filter>=23.0
Pillow>=10.0.0
requests>=2.31.0
python-decouple>=3.8
python-dotenv>=1.0.0